refactor: moved util components from data-upload to util folder
