+++ sites/geohub/src/components/util/CountryCard.test.ts
@@ -1,6 +1,6 @@
 import { describe, beforeEach, it, expect } from 'vitest';
 import { render, type RenderResult } from '@testing-library/svelte';
-import CountryCard from '$components/data-upload/CountryCard.svelte';
+import CountryCard from '$components/util/CountryCard.svelte';
 import type { Country } from '$lib/types';
 
 const exampleCountry: Country = {
