+++ sites/geohub/src/components/util/DataPreview.svelte
