+++ sites/geohub/src/components/util/DataPreviewContent.svelte
